moved everything to Toolbar.js, sort of almost fixed the Feed
